Specify a buffer size for EAC3 audio streams
There an issue in FireTV for assets with less than <5 seconds and with surround audio
tracks that causes the playback to freeze. This workaround the bug by providing a
buffer size, instead of allow AudioSink to determine it.
Default seems to be 196000 for the maximum bitrate of 6.4Mbit/sec for EAC3.
As I have never seen audio with bitrates above a 1 megabit, I think we should be safe
with lower numbers. For comparision, AC3 uses 20000 for buffer size.

Signed-off-by: Jorge Ruesga <jorge.ruesgacriado@sky.uk>