Media codec buffer workarounds
This PR brings the following workarounds:

- codecNeedsDrainWaitForEosStateWorkaround
- codecNeedsProcessedOutputBufferBeforeDrainWaitForEosStateWorkaround
- codecNeedsDrainWaitForEosStateAfterProcessOutputFormatWorkaround

----

Reinitialize EAC3 codec if on DRAIN_STATE_WAIT_END_OF_STREAM

Some codecs processes incorrectly drainAndReinitializeCodec and
renderer never get out from DRAIN_STATE_WAIT_END_OF_STREAM state, leading
audio and video to stall waiting for the codec to process the event,
so just force a codec reinitialization to get the codec in the
right state again.

This patch will only be applied to Amazon devices and only check
for drain status/action or input eos to determine whether to
apply or not the workarround.

----

Force E-AC3 codec reconfiguration and codecNeedsDrainWaitForEosStateWorkaround for ATV8 devices

This workaround aims to for some E-AC3 codecs whose incorrectly process
discontinuity events requiring a codec reconfiguration instead of just
a buffer flush.

The workaround is applied to:

 - Hisense AndroidTV 8.0 (laoshan)
 - Hisense AndroidTV 8.0 (xinhaoshan)
 - Xiami MiBox 3 Android TV 8.0 (once)

In addition, enable codecNeedsDrainWaitForEosStateWorkaround as well for the
following devices:

 - Hisense AndroidTV 8.0 (laoshan)
 - Hisense AndroidTV 8.0 (xinhaoshan)

----

Workaround TCL TV devices

This aims to fix issues TCVL TV devices (UnionTV) are suffering while playing eac3 assets. This devices
seems to suffer for codecNeedsDrainWaitForEosStateWorkaround when switching between main and ad content
and sometimes they also needs codecNeedsDrainWaitForEosStateAfterProcessOutputFormatWorkaround.

codecNeedsDrainWaitForEosStateAfterProcessOutputFormatWorkaround is a new workaround to force
codec to process EndOfStream, when previously the driver reported a INFO_OUTPUT_FORMAT_CHANGED, but
subsequently reported to have some output buffer to process, and inputStream was reported ended and
it is waiting for eos. In that case, the driver need to be force to process EOS to properly release
the resorces and not enter into an infinity loop, because even the codec reported something to
process, subsquence reads doesn't provide any data.

This patch is targeting only

SDK_INT: <= 28
MANUFACTURER: TCL
DEVICE: UnionTV
CODEC: OMX.RTK.audio.decoder

----

Signed-off-by: Jorge Ruesga <jorge.ruesgacriado@sky.uk>