Ignore zero period adaptation set
There is currently a bug where segments are stripped out from a period right before it
falls out of the live window. Since this is against DASH spec, it causes Exo to fail
playback. This method contains all of the regular MPD parsing found in
DashManifestParser, with a couple of extra lines to filter out periods without
adaptation sets.
This commit ensures that we don't orphan event streams that are still open and
valid - dependent on timing being based on PTS.

Signed-off-by: Jorge Ruesga <jorge.ruesgacriado@sky.uk>