Ad Resiliency - Detect stalled ad periods and notify throught AdMediaPeriod#maybeNotifyOrThrowMediaStreamError
This is workaround caused by HelioAdLoader#getSafeAdjustedSeekValue setting position at a negative
value for pre-roll, which in case of a ad manifest failure, it will cause ExoPlayer to enter into an
infinite loop and no reporting the ad manifest error. This PR just allow to advance to next period
(the ad period failed one) in order to get the manifest failure to be notified.

Basically, this Helio behaviour is causing AudioTrack to don't drain the input buffer, so input
buffer has always bytes to be processed. The AudioTrack logic is being for devices above API 21.
Since AudioTrack is platform specific there is no easy solution to bypass/correct this wrong behaviour.
The only solution seems to flush the buffer when the Audio renderer is stalled. This can be done by
force shouldAdvance to the next period and adjust the renderer position (which clears all renderers
buffer and check in the right period to raise the connection error dowload failure ad). That triggers
the AdMediaPeriod#maybeNotifyOrThrowMediaStreamError method which allow to buble up the ad insertion
failure.

The isSeekToHelioAdEdgeWorkaround workaround checks whether a render seek have happened at an advert
edge, considering the following conditions to avoid cause issues in other circunstances:

    * Only for AdMediaPeriod
    * When next MediaPeriodHolder is not prepared

Under that conditions, a renderer is stalled when:

    * The render position is negative and nextPlayingPeriodHolder starttime is 1us (pre-roll)
    * All rendererer have finished (hasReadStreamToEnd() == true and getReadingPositionUs() == C.TIME_END_OF_SOURCE)
    * There is a stalled renderer by more than 200ms (hasReadStreamToEnd() == false and
      getReadingPositionUs() != C.TIME_END_OF_SOURCE) and there are no other renderers not stalled and not ended.

Signed-off-by: Jorge Ruesga <jorge.ruesgacriado@sky.uk>