Do not bubble up onRenderedFirstFrame buffering events once it is loaded
https://github.com/sky-uk/core-video-team-exoplayer/commit/bc9a0860e#diff-e3030e4031c907d28d7f5195f6b107a2R853 aims to
ensures a more consistent playback behavior, but this commit also generates a an small buffering transition between
midrolls to main content in addition to between playlist. This is not a truly buffering event since, the media
codec rendered is really ready (the changes just forces a quicker render of the buffer), so there is no need to
notify clients about this "internal" change.

This commit traces when MediaCodecVideoRenderer rendered at least one first frame that doesn't require
a reset of the buffer and notify that is ready in that cases to avoid ExoPlayerImplInternal to change
ready status from READY -> BUFFERING -> READY because of the commit above mentioned.

Signed-off-by: Jorge Ruesga <jorge.ruesgacriado@sky.uk>