Restore windowStartTime notification for static manifests if availabilityStartTime is present
When transition from a dynamic to a static window, https://github.com/sky-uk/core-video-team-exoplayer/commit/01f57c3c0bc068a2ab66efcfec62e5b242df4903
remove the availability to notify window windowStartTime for static manifest. We rely on this
variable to calculate our seekable range.

This commits restores the notification of windowStartTime if the availabilityStartTime
information is present in the manifest.

Signed-off-by: Jorge Ruesga <jorge.ruesgacriado@sky.uk>